combined 2 contracts in one
